Tools: Documentation creator: Category update
Now, when exporting a page, the user will be prompted to select (or not) a category.
Each category will be listed in the index then the "No category" section.
Each category has a distinct subfolder.
Updated the css to make the documentation looks a bit nicer.
Fixed incorrect font.
Fixed incorrect webpage end when using nothing in the editor.